Move StepsMenu to components/ui, add Storybook
