--- airbyte-webapp/src/components/StepsMenu/index.tsx
@@ -1,4 +1,0 @@
-import StepsMenu from "./StepsMenu";
-
-export { StepsMenu };
-export default StepsMenu;
