+++ airbyte-webapp/src/components/ui/StepsMenu/Step.tsx
