+++ airbyte-webapp/src/components/ui/StepsMenu/index.tsx
@@ -1,0 +1,1 @@
+export * from "./StepsMenu";
