+++ airbyte-webapp/src/pages/ConnectionPage/pages/ConnectionItemPage/components/ConnectionPageTitle.tsx
@@ -5,7 +5,7 @@ import { useNavigate } from "react-router-dom";
 
 import { Text } from "components/base/Text";
 import { InfoBox } from "components/InfoBox";
-import StepsMenu from "components/StepsMenu";
+import { StepsMenu } from "components/ui/StepsMenu";
 
 import { ConnectionStatus, DestinationRead, SourceRead, WebBackendConnectionRead } from "core/request/AirbyteClient";
